Move php matcher related classes into a PhpMatcher namespace
