+++ src/PhpMatcher/MatcherFactory.php
@@ -1,5 +1,5 @@
 <?php declare(strict_types=1);
-namespace Behapi\Tools;
+namespace Behapi\PhpMatcher;
 
 use Coduo\PHPMatcher\Matcher;
 use Coduo\PHPMatcher\Factory\SimpleFactory;
@@ -9,20 +9,20 @@ use Coduo\PHPMatcher\Factory\SimpleFactory;
  *
  * @author Baptiste Clavié <clavie.b@gmail.com>
  */
-final class BehapiFactory extends SimpleFactory
+final class MatcherFactory extends SimpleFactory
 {
     /** @var Matcher\JsonMatcher */
     private $matcher;
 
+    public function __construct()
+    {
+        $orMatcher = $this->buildOrMatcher();
+        $this->matcher = new Matcher\JsonMatcher($orMatcher);
+    }
+
     /** {@inheritDoc} */
     public function createMatcher()
     {
-        if (null === $this->matcher) {
-            $orMatcher = $this->buildOrMatcher();
-
-            $this->matcher = new Matcher\JsonMatcher($orMatcher);
-        }
-
         return new Matcher($this->matcher);
     }
 }
